code/backend/requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
psycopg[binary]==3.2.3
alembic==1.13.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment & Configuration
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
email-validator==2.3.0

# Data Processing (ETL)
pandas==2.2.2
openpyxl==3.1.5
xlrd==2.0.1

# HTTP Client (for Gemini API)
httpx==0.27.2

# Development & Testing
pytest==8.3.3
pytest-asyncio==0.24.0
